+++ inst/scripts/RSienaSNADescriptives.R
@@ -2,7 +2,7 @@
 ###
 ### ---- RscriptSNADescriptives.R: a script for the introduction to RSiena -------
 ###
-###                               version: May 8, 2014
+###                               version: September 8, 2020
 ##################################################################################
 #
 # Rscript01DataFormat.R is followed by
@@ -33,23 +33,32 @@
 # For this script, you will need the data read and modified in the script
 # Rscript01DataFormat.R. If you have already ran that script, you may
 # load the required workspace:
-load("WorkspaceRscript01.RData")
+# load("WorkspaceRscript01.RData")
 
-        net1 <- as.network( friend.data.w1 )
-        net2 <- as.network( friend.data.w2 )
-        net3 <- as.network( friend.data.w3 )
-        plot.sociomatrix( net1,drawlab = F, diaglab = F, xlab = 'friendship t1' )
-        plot.sociomatrix( net2,drawlab = F, diaglab = F, xlab = 'friendship t2' )
-        plot.sociomatrix( net3,drawlab = F, diaglab = F, xlab = 'friendship t3' )
+# If not, to make this script self-contained, you may run the commands:
+
+    library(RSiena)
+    friend.data.w1 <- s501
+    friend.data.w2 <- s502
+    friend.data.w3 <- s503
+    drink <- s50a
+    smoke <- s50s
+
+    net1 <- as.network( friend.data.w1 )
+    net2 <- as.network( friend.data.w2 )
+    net3 <- as.network( friend.data.w3 )
+    plot.sociomatrix( net1,drawlab = F, diaglab = F, xlab = 'friendship t1' )
+    plot.sociomatrix( net2,drawlab = F, diaglab = F, xlab = 'friendship t2' )
+    plot.sociomatrix( net3,drawlab = F, diaglab = F, xlab = 'friendship t3' )
 
 # The class,
-	class( net1 )
+    class( net1 )
 # with attributes
-	attributes( net1 )
+    attributes( net1 )
 # has special methods associated with it.
 # while  plot( friend.data.w1 ) only produces a rather dull plot of
 # the first two columns
-#       plot( net1, xlab = 'friendship t1' )
+#     plot( net1, xlab = 'friendship t1' )
 # produces a nice sociogram
 #
 # Some further descriptives you can do for the data are plotting and
@@ -57,59 +66,59 @@ load("WorkspaceRscript01.RData")
 
 # add the attribute drink to the network object
 
-        net1 %v% "drink" <- drink[ , 1 ]
+    net1 %v% "drink" <- drink[ , 1 ]
 
 # color the nodes by drink
 
-        plot( net1, vertex.col = "drink", xlab = 'friendship t1' )
+    plot( net1, vertex.col = "drink", xlab = 'friendship t1' )
 
 
 # Now let's color the nodes by drink and scale the vertex by degree of nodes!
 #
 # First calculate degree:
 
-        deg <- rowSums( as.matrix( net1 ) )# NB:  rowSums() is defined for class matrix
+    deg <- rowSums( as.matrix( net1 ) )# NB:  rowSums() is defined for class matrix
 
 # have a look at the degree distribution
 
-        table( deg, useNA = 'always' )
+    table( deg, useNA = 'always' )
 
 # Now do the desired plot:
 
-        plot( net1, vertex.col = "drink", vertex.cex = (deg + 1)/1.5 )
+    plot( net1, vertex.col = "drink", vertex.cex = (deg + 1)/1.5 )
 
 # ---- Plot the three waves of data --------------------------------------------
 
 # Add drink to waves 2 and 3
-        net2 %v% "drink" <- drink[ , 2 ]
-        net3 %v% "drink" <- drink[ , 3 ]
-      	deg2 <- rowSums( as.matrix( net2 ) )
-      	deg3 <- rowSums( as.matrix( net3 ) )
+    net2 %v% "drink" <- drink[ , 2 ]
+    net3 %v% "drink" <- drink[ , 3 ]
+    deg2 <- rowSums( as.matrix( net2 ) )
+    deg3 <- rowSums( as.matrix( net3 ) )
 
 # Create a set of panels ( 1 row by 3 columns, or 3 columns by 1 row)
 
-        par( mfrow = c( 1, 3 ) )
+    par( mfrow = c( 1, 3 ) )
 
 # creating three plots after each other will place them in consecutive panels
 
-        plot( net1, vertex.col = "drink", vertex.cex = (deg + 1)/1.5 )
-        plot( net2, vertex.col = "drink", vertex.cex = (deg2 + 1)/1.5 )
-        plot( net3, vertex.col = "drink", vertex.cex = (deg3 + 1)/1.5 )
+    plot( net1, vertex.col = "drink", vertex.cex = (deg + 1)/1.5 )
+    plot( net2, vertex.col = "drink", vertex.cex = (deg2 + 1)/1.5 )
+    plot( net3, vertex.col = "drink", vertex.cex = (deg3 + 1)/1.5 )
 
 # Each time we make a plot the coordinates move - because always
 # the starting values are random. We can also save coordinates
 # and use them for later plotting:
 
-        par( mfrow = c( 1, 3 ) )
-        coordin <-  plot( net1, vertex.col = "drink", vertex.cex = (deg +1 )/1.5 )
-        plot( net2, coord = coordin, vertex.col = "drink", vertex.cex = (deg2 + 1)/1.5 )
-        plot( net3, coord = coordin, vertex.col = "drink", vertex.cex = (deg3 + 1) /1.5 )
+    par( mfrow = c( 1, 3 ) )
+    coordin <-  plot( net1, vertex.col = "drink", vertex.cex = (deg +1 )/1.5 )
+    plot( net2, coord = coordin, vertex.col = "drink", vertex.cex = (deg2 + 1)/1.5 )
+    plot( net3, coord = coordin, vertex.col = "drink", vertex.cex = (deg3 + 1) /1.5 )
 
 # To get coordinates based on all three waves: coordin <-  plot( net1 + net2 + net3 )
 # For more plotting options, try the gplot function in the "sna" library
 
-        ?gplot
-        ?gplot.layout
+    ?gplot
+    ?gplot.layout
 
 
 # ---- Basic network statistics ------------------------------------------------
@@ -118,37 +127,53 @@ load("WorkspaceRscript01.RData")
 # The following are examples.
 # some important graph level statistics
 
-        gden( net1 ) # density
-        grecip( net1 ) # proportion of dyads that are symmetric
-        grecip( net1, measure = "dyadic.nonnull" ) # reciprocity, ignoring the null dyads
-        gtrans( net1 ) # transitivity
+    gden( net1 ) # density
+    grecip( net1 ) # proportion of dyads that are symmetric
+    grecip( net1, measure = "dyadic.nonnull" ) # reciprocity, ignoring the null dyads
+    gtrans( net1 ) # transitivity
 
 # dyad and triad census
 
-        dyad.census( net1 )
-        triad.census( net1 )
+    dyad.census( net1 )
+    triad.census( net1 )
 
 # out degree distribution (of course for a symmetric network outdegree=indegree)
 
-        outdegree <- degree( net1, cmode = "outdegree" )
-        outdegree #outgoing ties of each note
+    outdegree <- degree( net1, cmode = "outdegree" )
+    outdegree #outgoing ties of each note
 
-        hist( outdegree )
-        quantile( outdegree )
+    hist( outdegree )
+    quantile( outdegree )
 
 # measures of connectivity and distance
 
-        dist <- geodist(net1, inf.replace = Inf, count.paths = TRUE)
+    dist <- geodist(net1, inf.replace = Inf, count.paths = TRUE)
 # calculate the geodesic distance (shortest path length) matrix
-   	   dist$gd
+    dist$gd
 # matrix of geodesic distances
-	      dist$counts
-# matrix containing the number of paths at each geodesic between each pair of vertices
-        dist
-        reach <- reachability( net1 )  # calculate the reachability matrix
-        reach
+    dist$counts
+    table(dist$counts)
+# reachability matrix:
+    ?reachability
+    reach <- reachability( net1 )  # calculate the reachability matrix
+    reach
+
+
+# ---- Network autocorrelation  ------------------------------------------------
+
+# Moran's autocorrelation for outgoing ties:
+
+nacf(net1, drink[, 1], type="moran", neighborhood.type='out')[2]
+nacf(net2, drink[, 2], type="moran", neighborhood.type='out')[2]
+nacf(net3, drink[, 3], type="moran", neighborhood.type='out')[2]
+
+# Moran's autocorrelation for outgoing and incoming ties:
+nacf(net1, drink[, 1], type="moran", neighborhood.type='total')[2]
+nacf(net2, drink[, 2], type="moran", neighborhood.type='total')[2]
+nacf(net3, drink[, 3], type="moran", neighborhood.type='total')[2]
+
 ################################################################################
 ###
-### ---- PROCEED TO Rscript02SienaVariableFormat.R FOR PREPARING DATA FOR RSIENA -
+### -- PROCEED TO Rscript02SienaVariableFormat.R FOR PREPARING DATA FOR RSIENA -
 ###
 ################################################################################